backend/pyproject.toml
# backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "llmcoder-backend"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn",
    "SQLAlchemy",
    "python-dotenv",
    "passlib",
    "jose",
    "requests",
    "atlassian-python-api",
    "google-api-python-client",
    "tweepy",
    "newsapi-python",
    "plaid-python",
    "beautifulsoup4",
    "lxml",
    "playwright",
    "nest-asyncio",
     "python-multipart",
     "httpx",
    "scrapy",
    "sentence-transformers",
    "openai",
    "python-apt"
]